Add a gir include dependency graph tracking class
This is required because we want to parse a gir file that defines
a type before parsing those who use that type. This means the
gir files to be parsed must be topologically sorted in reverse
order based on the include dependency graph before parsing begins.
The main role of the new GirDependencyGraph class is exactly this
sorting.