Improve type reference collection using TypeRegistry
Now that normal parsing creates a TypeRegistry from all the defined
types it finds before starting dumb type reference collection,
we can use it to tell if a type reference is of a global type, or
a locally defined one, without any guesswork from the type name.
The previous "heuristic" method tried to guess this from the
format of the name, but it did not work in all cases.