Prefix generated class names
Using only the class name declared in a gir file as the name of
the corresponding generated Java class can lead to confusing
results. An example is having three different classes called
Object: the classes corresponding to GObject and GstObject are
both called Object, with the additional java.lang.Object making
things even worse.

To solve this, prefix the names of all generated classes (except
the namespace-wide function classes) with the c:identifier-prefixes
of the namespace, or if not present, the name of the namespace.