Prefix record/class field getters and setters with field_
Some method names in .gir files conflict with field names in their
classes or records. With the current generation scheme, both methods,
and fields are bound to methods in their classes, named after the
name declared in the .gir files. This may lead to the generation of
multiple methods with the same signature in the same class. Solve
this by prefixing field getters and setters with field_.