Make generated methods without an instance-parameter static
This is still not a perfect way of deciding what should be static
and what not. It results in things like GList.nth_data(list) instead
of the sensible list.nth_data() because while list here is obviously
an instance parameter, it is not declared so in the gir file. But,
rather have this than GList.alloc() not being static, or everything
being static, I think.