Make private fields private in the generated code as well
They cause name clashes when a class and its superclass both declare
a pointer field named "priv", but the pointers are differently typed.
We do not have to know about private fields anyways, so make them
private to prevent inheritance. Leave them in place so that BridJ
knows about them.