Do not make private class/record member getters/setters private
BridJ does not count the fields this way, which alters the size of the
enclosing struct type as seen by BridJ. This causes problems with nested
structures, which is used extensively by GStreamer. (The nested structure
is believed by BridJ to be shorter than it actually is, therefore
potentially breaking offset calculation for other fields.)

Since the getters/setters are prefixed with the class name, no
inheritance-related name conflicts should occur, therefore it is safe to
make them public again.