Do not generate duplicate methods
There are cases where multiple methods are indicated in the gir files
to map to the same native function. If these are generated into the same
class, the generated code breaks. An example is when a non-virtual method
is defined globally, and also in an interface. Both are generated into the
namespace-wide catch-all class. This commit prevents the generation of these
duplicates.